+++ Adj/Program/Progress.hs
@@ -1,13 +1,13 @@
 module Adj.Program.Progress where
 
-import Adj.Algebra.Category ((:+:) (This, That))
+import Adj.Algebra.Category (Flat (Flat), (:+:) (This, That))
 
-type Progress stop = (:+:) stop
+type Progress stop = Flat (:+:) stop
 
 pattern Continue :: continue -> Progress stop continue
-pattern Continue continue <- That continue
-	where Continue continue = That continue
+pattern Continue continue <- Flat (That continue)
+	where Continue continue = Flat (That continue)
 
 pattern Stop :: stop -> Progress stop continue
-pattern Stop stop <- This stop
-	where Stop stop = This stop
+pattern Stop stop <- Flat (This stop)
+	where Stop stop = Flat (This stop)